Replaced react tooltip from navigation bar with PageflowTooltip
