+++ entry_types/scrolled/package/src/frontend/navigation/ChapterLink.module.css
@@ -25,3 +25,19 @@
     display: block;
   }
 }
+
+.chapterLinkTooltip {
+  border-bottom: 3px solid #e10028;
+  width: 200px;
+}
+
+.tooltipHeadline {
+  margin: 5px 0 0px;
+  color: #e10028;
+}
+
+@media (max-width: 780px) {
+  .chapterLinkTooltip {
+    display: none !important;
+  }
+}
