--- entry_types/scrolled/package/src/frontend/navigation/ChapterLinkTooltip.module.css
@@ -1,15 +1,0 @@
-.chapterLinkTooltip {
-  border-bottom: 3px solid #e10028;
-  width: 200px;
-}
-
-.tooltipHeadline {
-  margin: 5px 0 0px;
-  color: #e10028;
-}
-
-@media (max-width: 780px) {
-  .chapterLinkTooltip {
-    display: none !important;
-  }
-}
