--- entry_types/scrolled/package/src/frontend/navigation/LegalInfoTooltip.module.css
@@ -1,17 +1,0 @@
-.legalInfoTooltip {
-  width: 200px;
-  max-width: 200px;
-  text-align: left;
-}
-
-.legalInfoTooltip:after {
-  left: 90% !important;
-}
-
-.legalInfoTooltip p {
-  margin: 0 0 0.5em;
-}
-
-.legalInfoTooltip a {
-  color: #e10028;
-}
