+++ entry_types/scrolled/package/src/frontend/navigation/SharingMenu.module.css
@@ -3,3 +3,46 @@
   height: 40px;
   margin: 5px 0px;
 }
+
+:global(header .share svg) {
+  fill: rgb(0, 55, 90);
+}
+
+:global(header .share:hover svg) {
+  fill: #e10028;
+}
+
+.sharingTooltip {
+  width: 160px;
+  padding: 0 !important;
+}
+
+.sharingTooltip:after {
+  left: 90% !important;
+}
+
+.shareLinkContainer {
+  display: inline-block;
+  width: 80px;
+  height: 60px;
+  cursor: pointer;
+  color: transparent;
+  text-align: center;
+}
+
+.shareLink {
+  position: relative;
+  color: rgb(0, 55, 90);
+  text-decoration: none;
+}
+
+.shareLink:hover {
+  color: #e10028;
+}
+
+.shareIcon {
+  width: 80px;
+  height: 25px;
+  margin-top: 5px;
+  margin-bottom: 3px;
+}
